requirements.txt
# OpenTelemetry packages
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation==0.42b0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.42b0

# MarkLogic packages
marklogic-python-client>=1.1.0
requests>=2.31.0
python-dotenv>=1.0.0 